ajustes na semantica e na organizacao
