+++ src/components/Button/index.jsx
@@ -1,6 +1,5 @@
-import styles from "./Button.module.css";
+import styles from "./style.module.css";
 function Button() {
-
   return (
     <div className={styles.DivBtn}>
       <button type="button" className={styles.btn}></button>
