+++ src/components/Input/InputEmail.jsx
@@ -2,7 +2,7 @@ import styles from "./Input.module.css";
 function InputEmail() {
   return (
     <div className={styles.DivInput}>
-      <input type="email" placeholder='E-MAIL'className={styles.Input}></input>
+      <input type="email" placeholder="E-MAIL"className={styles.Input}></input>
     </div>
   );
 }
