+++ src/components/Input/InputPassword.jsx
@@ -2,7 +2,7 @@ import styles from "./Input.module.css"
 function InputPassword(){
   return(
    <div className={styles.DivInput}>
-     <input type="password" className={styles.Input}placeholder='SENHA'></input>
+     <input type="password" className={styles.Input}placeholder="SENHA"></input>
    </div>
   )
  }
